# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synapse Business Suite API",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/orquestador/start": {
      "post": {
        "tags": [
          "Orquestador"
        ],
        "summary": "Start Orquestador",
        "operationId": "start_orquestador_api_orquestador_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationTask"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orquestador/chat": {
      "post": {
        "tags": [
          "Orquestador"
        ],
        "summary": "Chat With Orquestador",
        "operationId": "chat_with_orquestador_api_orquestador_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AutomationTask": {
        "properties": {
          "TriggerID": {
            "type": "integer",
            "title": "Triggerid"
          },
          "ActionCommand": {
            "type": "string",
            "title": "Actioncommand"
          },
          "IsActive": {
            "type": "boolean",
            "title": "Isactive"
          },
          "LastTriggered": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasttriggered"
          }
        },
        "type": "object",
        "required": [
          "TriggerID",
          "ActionCommand",
          "IsActive"
        ],
        "title": "AutomationTask"
      },
...
      "ChatMessage": {
        "properties": {
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "message"
        ],
        "title": "ChatMessage"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}